+++ R/fitstats.R
@@ -444,7 +444,7 @@ fitstat_register = function(type, fun, alias = NULL, subtypes = NULL){
 #'
 #' @param x A `fixest` estimation, obtained for example from [`feols`].
 #' @param type Character vector or one sided formula. No default. 
-#' The type of fit statistic to be computed. 
+#' The type of fit statistic or tests to be computed. 
 #' The classic ones are: `n`, `rmse`, `r2`, `pr2`, `f`, `wald`, `ivf`, `ivwald`. 
 #' You have the full list in the details section or use `show_types = TRUE`. 
 #' Further, you can register your own types with [`fitstat_register`].
@@ -456,10 +456,15 @@ fitstat_register = function(type, fun, alias = NULL, subtypes = NULL){
 #' is returned instead (i.e. the same object without the class).
 #' @param show_types Logical, default is `FALSE`. If `TRUE`, this prompts all available types
 #' and nothing is returned.
+#' @param htest Logical scalar, default is `FALSE`. If `TRUE`, the results of tests are formatted
+#' according to the `htest` class (defined in the `stats` package), 
+#' and the returned object is a simple list. 
+#' If `FALSE`, the results of each test is a simple list but the object returned is of class 
+#' `fixest_fitstat`. 
 #' @param frame An environment in which to evaluate variables, default is `parent.frame()`. 
 #' Only used if the argument `type` is a formula and some values in the formula have to be 
-#' extended with the dot square bracket operator. Mostly for internal use.
-#' @param ... For internal use.
+#' extended with the dot square bracket operator. Mostly for internal use. 
+#' @param ... For internal use. 
 #' 
 #' @details 
 #' 
@@ -543,32 +548,52 @@ fitstat_register = function(type, fun, alias = NULL, subtypes = NULL){
 #'
 #'
 #' @return
-#' By default an object of class `fixest_fitstat` is returned. This object is a simple list 
-#' containing the statistics/tests requested by the user. For example `fitstat(x, c("r2", "f"))`
-#' returns a list with two elements named `r2` and `f`.
+#' 
+#' By default an object of class `fixest_fitstat` is returned. 
+#' This object is a simple list containing the statistics/tests requested 
+#' by the user. 
+#' For example `fitstat(x, c("r2", "f"))` returns a list with two elements named `r2` and `f`.
 #' 
 #' Each statistic from the `fixest_fitstat` object can be of two types:
 #' \enumerate{
-#' \item a list of numeric scalars whose elements depend on the type of fit statistic/test. 
-#' For example the F-test, accessed with `f`, contains the elements 
-#' `stat`, `p`, `df1`, and `df2`. The `wald` test contains the elements 
-#' `stat`, `p`, `df1`, `df2` and `vcov`. The likelihood ratio, `lr`, contains the elements
-#' `stat`, `p`, `df`. Etc.
+#'   \item a list of numeric scalars whose elements depend on the type of fit statistic/test. 
+#'   For example the F-test, accessed with `f`, contains the elements 
+#'   `stat`, `p`, `df1`, and `df2`. The `wald` test contains the elements 
+#'   `stat`, `p`, `df1`, `df2` and `vcov`. The likelihood ratio, `lr`, contains the elements
+#'   `stat`, `p`, `df`. Etc.
 #' 
-#' \item a numeric scalar, for either: i) scalar fit statistics, like `r2` or `rmse`, 
-#' ii) when a specific element of a fit statistic is accessed, like e.g. `f.stat` 
-#' which reports the `stat` element of the fit statistic `f`.
+#'   \item a numeric scalar, for either: i) scalar fit statistics, like `r2` or `rmse`, 
+#'   ii) when a specific element of a fit statistic is accessed, like e.g. `f.stat` 
+#'   which reports the `stat` element of the fit statistic `f`.
 #' }
 #' 
 #' The types of fit statistics and their structure are detailed 
 #' in the section `Available types` of this documentation.
 #' 
+#' The class `fixest_fitstat` has only a dedicated print method.
+#' 
+#' If the argument `htest=TRUE`, the object returned is a plain list containing the requested
+#' statistics and each test is formatted according to the `htest` class from the `stats` package. 
+#' `htest` objects are lists containing the following elements:
+#' \describe{
+#'   \item{statistic}{A numeric scalar equal to test statistic.}
+#'   \item{p.value}{A numeric scalar equal to the p-value of the test.}
+#'   \item{parameter}{A list containing various parameters used to calculate the 
+#'     p-value and the statistic. It can be `df`, `df1`, `df2`,  or `vcov`.}
+#'   \item{alternative}{The alternative hypothesis.}
+#'   \item{method}{The name of the test.}
+#'   \item{data.name}{The estimation call on which the test is applied.}
+#' }
+#' 
 #' Using `verbose = FALSE` removes the `fixest_fitstat` class from the returned object,
-#' turning it into a plain list.
+#' turning it into a plain list (this is ignored when `htest=TRUE`).
 #' 
 #' If only one type is selected, `simplify = TRUE` leads to the selected statistic to 
 #' be directly returned (and hence is not nested inside a list). 
 #' For example `fitstat(x, "r2", simplify = TRUE)` returns a simple scalar.
+#' 
+#' If `show_types=TRUE`, this function returns instead a character vector with
+#'  all the available types.
 #'
 #' @examples
 #'
@@ -586,6 +611,9 @@ fitstat_register = function(type, fun, alias = NULL, subtypes = NULL){
 #'
 #' # For wald and wf, you could show the pvalue instead:
 #' etable(gravity, fitstat = ~ rmse + r2 + wald.p + wf.p)
+#' 
+#' # You can display the tests with the htest format with htest=TRUE
+#' fitstat(gravity, ~ rmse + r2 + wald + wf, htest = TRUE)
 #'
 #' # Now let's display some statistics that are not built-in
 #' # => we use fitstat_register to create them
@@ -604,7 +632,7 @@ fitstat_register = function(type, fun, alias = NULL, subtypes = NULL){
 #'
 fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL, 
                    simplify = FALSE, verbose = TRUE, show_types = FALSE,
-                   frame = parent.frame(), ...){
+                   htest = FALSE, frame = parent.frame(), ...){
 
   r2_types = c("sq.cor", "cor2", "r2", "ar2", "pr2", "apr2", "par2", "wr2",
                "war2", "awr2", "wpr2", "pwr2", "wapr2", "wpar2", "awpr2",
@@ -687,7 +715,7 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
 
   check_arg(x, "class(fixest) mbt")
   check_set_arg(type, "character vector no na | os formula")
-  check_arg(simplify, verbose, "logical scalar")
+  check_arg(simplify, verbose, htest, "logical scalar")
 
   if(inherits(type, "formula")){
     type = .xpd(type, frame = frame)
@@ -699,6 +727,13 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
   # To update
   if(!isTRUE(x$summary)){
     x = summary(x, vcov = vcov, cluster = cluster, ssc = ssc, warn = FALSE)
+  }
+  
+  if(htest){
+    # we update the type dict (can be modifed by the user)
+    all_types = fitstat(give_types = TRUE)
+    dict_type = all_types$R_alias
+    options(fixest_fitstat_type_R_alias = dict_type)
   }
 
   IS_ETABLE = isTRUE(dots$etable)
@@ -726,6 +761,7 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
 
   for(i in seq_along(type_all)){
     type = type_all[i]
+    alternative = NULL
     
     # Big if
     if(is_NA_model){
@@ -807,6 +843,13 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
 
       if(root == "f"){
         if(!is.null(x$ssr)){
+          
+          if("fixef_id" %in% names(x)){
+            alternative = "At least one coefficient (including the fixed-effects) is different from 0"
+          } else {
+            alternative = "At least one coefficient is different from 0"
+          }
+          
           df1 = degrees_freedom(x, "k") - 1
           df2 = degrees_freedom(x, "t")
 
@@ -828,6 +871,9 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
           res_all[[type]] = NA
         }
       } else if(root == "wf"){
+        
+        alternative = "At least one non fixed-effect coefficient is different from 0"
+        
         df1 = length(x$coefficients)
 
         if(!is.null(x$ssr_fe_only) && df1 > 0){
@@ -845,6 +891,9 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root == "ivfall"){
         if(isTRUE(x$is_iv)){
           if(x$iv_stage == 1){
+            
+            alternative = "At least one coefficient associated to the instruments is different from 0"
+            
             df1 = degrees_freedom(x, vars = x$iv_inst_names_xpd)
             df2 = degrees_freedom(x, "resid")
 
@@ -855,6 +904,8 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
 
           } else {
             # f stat for the second stage
+            
+            alternative = "At least one coefficient associated to the endogenous regressors is different from 0"
 
             df1 = degrees_freedom(x, vars = x$iv_endo_names_fit)
             df2 = degrees_freedom(x, "resid")
@@ -876,6 +927,9 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root == "ivf1"){
 
         if(isTRUE(x$is_iv)){
+          
+          alternative = "At least one coefficient associated to the instruments is different from 0"
+          
           df1 = degrees_freedom(x, vars = x$iv_inst_names_xpd, stage = 1)
           df2 = degrees_freedom(x, "resid", stage = 1)
 
@@ -895,6 +949,16 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
               vec = list(stat = stat, p = p, df1 = df1, df2 = df2)
               res_all[[paste0(type, "::", endo)]] = set_value(vec, value)
             }
+            
+            if(htest){
+              # special case
+              for(endo in names(x_first)){
+                type_endo = paste0(type, "::", endo)
+                res_all[[type_endo]] = fitstat_to_htest(x, type, res_all[[type_endo]], alternative)
+              }
+              
+              next
+            }
           }
 
         } else {
@@ -904,6 +968,7 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root == "ivf2"){
         if(isTRUE(x$is_iv) && x$iv_stage == 2){
           # f stat for the second stage
+          alternative = "At least one coefficient associated to the endogenous regressors is different from 0"
 
           df1 = degrees_freedom(x, vars = x$iv_endo_names_fit)
           df2 = degrees_freedom(x, "resid")
@@ -924,6 +989,7 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root == "kpr"){
         if(isTRUE(x$is_iv) && x$iv_stage == 2){
           # The KP rank test is computed in a specific function
+          alternative = "The instruments are strong"
 
           vec = kp_stat(x)
           res_all[[type]] = set_value(vec, value)
@@ -935,6 +1001,7 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root == "cd"){
         if(isTRUE(x$is_iv) && x$iv_stage == 2){
           # The KP rank test is computed in a specific function
+          alternative = "The instruments are strong"
 
           vec = cd_stat(x)
           res_all[[type]] = set_value(vec, value)
@@ -946,6 +1013,12 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root == "wald"){
         # Joint nullity of the coefficients
         # if FE => on the projected model only
+        
+        if("fixef_id" %in% names(x)){
+          alternative = "At least one coefficient is different from 0"
+        } else {
+          alternative = "At least one non fixed-effect coefficient is different from 0"
+        }
 
         qui = !names(x$coefficients) %in% "(Intercept)"
         my_coef = x$coefficients[qui]
@@ -970,6 +1043,8 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
 
         if(isTRUE(x$is_iv)){
           if(x$iv_stage == 1){
+            
+            alternative = "At least one coefficient associated to the instruments is different from 0"
 
             df1 = degrees_freedom(x, vars = x$iv_inst_names_xpd)
             df2 = degrees_freedom(x, "resid")
@@ -983,6 +1058,8 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
 
           } else {
             # wald stat for the second stage
+            
+            alternative = "At least one coefficient associated to endogenous regressor is different from 0"
 
             df1 = degrees_freedom(x, vars = x$iv_endo_names_fit)
             df2 = degrees_freedom(x, "resid")
@@ -1002,9 +1079,11 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root %in% "ivwald1"){
 
         if(isTRUE(x$is_iv)){
+          
+          alternative = "At least one coefficient associated to the instruments is different from 0"
+          
           df1 = degrees_freedom(x, vars = x$iv_inst_names_xpd, stage = 1)
           df2 = degrees_freedom(x, "resid", stage = 1)
-
 
           if(x$iv_stage == 1){
 
@@ -1042,6 +1121,17 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
                          vcov = attr(x$cov.scaled, "vcov_type"))
               res_all[[paste0(type, "::", endo)]] = set_value(vec, value)
             }
+            
+            if(htest){
+              # special case
+              for(endo in names(x_first)){
+                type_endo = paste0(type, "::", endo)
+                res_all[[type_endo]] = fitstat_to_htest(x, type, res_all[[type_endo]], alternative)
+              }
+              
+              next
+            }
+            
           }
 
         } else {
@@ -1051,6 +1141,7 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root == "ivwald2"){
         if(isTRUE(x$is_iv) && x$iv_stage == 2){
           # wald stat for the second stage
+          alternative = "At least one coefficient associated to the endogenous regressors is different from 0"
 
           df1 = degrees_freedom(x, vars = x$iv_endo_names_fit)
           df2 = degrees_freedom(x, "resid")
@@ -1069,6 +1160,8 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root == "wh"){
         if(isTRUE(x$is_iv) && x$iv_stage == 2){
           # Wu Hausman stat for the second stage
+          alternative = "The regressor is endogenous"
+          
           vec = x$iv_wh
           res_all[[type]] = set_value(vec, value)
 
@@ -1079,6 +1172,8 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
       } else if(root == "sargan"){
         if(!is.null(x$iv_sargan)){
           # Wu Hausman stat for the second stage
+          alternative = "At least one instrument is invalid"
+          
           vec = x$iv_sargan
           res_all[[type]] = set_value(vec, value)
 
@@ -1088,6 +1183,7 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
 
       } else if(root == "lr"){
         if(!is.null(x$ll_null)){
+          alternative = "At least one coefficient is different from 0"
 
           stat = 2 * (x$loglik - x$ll_null)
           df = x$nparams
@@ -1102,7 +1198,7 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
 
       } else if(root == "wlr"){
         if(!is.null(x$ll_fe_only) || (x$method_type == "feglm" && !is.null(x$fixef_id))){
-
+          alternative = "At least one of the non fixed-effect coefficients is different from 0"
           if(x$method_type == "feglm"){
             # estimation of the FE only model
 
@@ -1137,6 +1233,10 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
         stop("The type '", type, "' is not supported, see details of ?fitstat, or use fitstat(show_types = TRUE) to get the names of all supported tests.")
       }
     }
+    
+    if(htest){
+      res_all[[type]] = fitstat_to_htest(x, type, res_all[[type]], alternative)
+    }
 
   }
 
@@ -1144,7 +1244,7 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
     verbose = FALSE
   }
 
-  if(verbose){
+  if(verbose && !htest){
     class(res_all) = "fixest_fitstat"
   }
 
@@ -1153,6 +1253,48 @@ fitstat = function(x, type, vcov = NULL, cluster = NULL, ssc = NULL,
   }
 
   res_all
+}
+
+
+fitstat_to_htest = function(est, type, fitstat_test, alternative){
+  
+  if(!is.list(fitstat_test)){
+    return(fitstat_test)
+  }
+  
+  my_test = list()
+  if("stat" %in% names(fitstat_test)){
+    my_test$statistic = setNames(fitstat_test$stat, "statistic")
+  }
+  
+  if("p" %in% names(fitstat_test)){
+    my_test$p.value = fitstat_test$p
+  }
+  
+  prm_names = setdiff(names(fitstat_test), c("stat", "p"))
+  if(length(prm_names) > 0){
+    prms = fitstat_test[prm_names]
+    my_test$parameter = prms
+  }
+  
+  dict_type = getOption("fixest_fitstat_type_R_alias")
+  if(is.null(dict_type)){
+    # we make sure it's OK and it works, even when not called in the right sequence
+    all_types = fitstat(give_types = TRUE)
+    dict_type = all_types$R_alias
+    options(fixest_fitstat_type_R_alias = dict_type)
+  }
+  
+  if(!is.null(alternative)){
+    my_test$alternative = alternative
+  }
+  
+  my_test$method = if(type %in% names(dict_type)) dict_type[[type]] else type
+  my_test$data.name	= paste(deparse(est$call, width.cutoff = 100, nlines = 3), collapse = "\n")
+  
+  class(my_test) = "htest"
+  
+  my_test
 }